clientApp/requirements.txt
requests>=2.26.0
simplejson>=3.16.0
python-dateutil>=2.7.3